Separate creation of notification to a new class
