+++ app/src/main/java/ch/epfl/polychef/notifications/NotificationReceiverService.java
@@ -1,16 +1,10 @@
 package ch.epfl.polychef.notifications;
 
-import android.app.NotificationChannel;
-import android.app.NotificationManager;
-import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
-import android.os.Build;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.RequiresApi;
-import androidx.core.app.NotificationCompat;
 
 import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.messaging.FirebaseMessagingService;
@@ -18,11 +12,9 @@ import com.google.firebase.messaging.RemoteMessage;
 
 import java.util.List;
 import java.util.Objects;
-import java.util.Random;
 
 import ch.epfl.polychef.CallHandler;
 import ch.epfl.polychef.MultipleCallHandler;
-import ch.epfl.polychef.R;
 import ch.epfl.polychef.pages.HomePage;
 import ch.epfl.polychef.recipe.Recipe;
 import ch.epfl.polychef.recipe.RecipeStorage;
@@ -31,7 +23,6 @@ import ch.epfl.polychef.users.UserStorage;
 
 public class NotificationReceiverService extends FirebaseMessagingService {
     private static final String TAG = "MessagingService";
-    private String channelID = "User notification";
 
     public NotificationReceiverService() {
     }
@@ -61,7 +52,7 @@ public class NotificationReceiverService extends FirebaseMessagingService {
                 public void onSuccess(Recipe data) {
                     Intent intent = new Intent(NotificationReceiverService.this.getContext(), HomePage.class);
                     intent.putExtra("RecipeToSend", data);
-                    setNotificationWithIntent(remoteMessage, intent);
+                    getNotificationUtils().setNotificationWithIntent(getContext(), remoteMessage, intent);
                 }
 
                 @Override
@@ -71,37 +62,9 @@ public class NotificationReceiverService extends FirebaseMessagingService {
             });
 
         } else {
-            setNotificationWithIntent(remoteMessage, new Intent(getContext(), HomePage.class));
+            getNotificationUtils().setNotificationWithIntent(getContext(), remoteMessage, new Intent(getContext(), HomePage.class));
         }
 
-    }
-
-    private void setNotificationWithIntent(RemoteMessage remoteMessage, Intent intent) {
-        NotificationManager notificationManager = (NotificationManager)getContext().getSystemService(Context.NOTIFICATION_SERVICE);
-        int notificationID = getRandom();
-        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-        PendingIntent pendingIntent = PendingIntent.getActivity(getContext() , 0, intent,
-                PendingIntent.FLAG_ONE_SHOT);
-        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
-            setChannels(notificationManager);
-        }
-        NotificationCompat.Builder notificationBuilder = new NotificationCompat.Builder(getContext(), channelID)
-                .setSmallIcon(R.drawable.ic_star_black_yellow)
-                .setContentTitle(remoteMessage.getData().get("title"))
-                .setContentText(remoteMessage.getData().get("message"))
-                .setAutoCancel(true)
-                .setContentIntent(pendingIntent);
-        notificationManager.notify(notificationID, notificationBuilder.build());
-    }
-
-    @RequiresApi(api = Build.VERSION_CODES.O)
-    private void setChannels(NotificationManager notificationManager){
-        NotificationChannel adminChannel;
-        adminChannel = new NotificationChannel(channelID, "New notification", NotificationManager.IMPORTANCE_HIGH);
-        adminChannel.setDescription("New user notification");
-        if (notificationManager != null) {
-            notificationManager.createNotificationChannel(adminChannel);
-        }
     }
 
     public RecipeStorage getRecipeStorage() {
@@ -116,11 +79,11 @@ public class NotificationReceiverService extends FirebaseMessagingService {
         return FirebaseMessaging.getInstance();
     }
 
-    public int getRandom() {
-        return new Random().nextInt(3000);
-    }
-
     public Context getContext() {
         return this;
+    }
+
+    public NotificationUtils getNotificationUtils() {
+        return NotificationUtils.getInstance();
     }
 }
